Baselines/16-20/Experiment/requirements.txt
cog==0.15.1
decord==0.6.0
fairscale==0.4.13
numpy==2.2.6
Pillow==11.2.1
pycocoevalcap==1.2
pycocotools==2.0.9
ruamel_yaml==0.18.12
timm==1.0.15
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.52.4
